exxtrat population initialisation and add main class
