+++ microsim/utilities.py
@@ -2,6 +2,8 @@
 import pandas as pd
 from typing import List
 
+from microsim.column_names import ColumnNames
+
 class Optimise:
     """
     Functions to optimise the memory use of pandas dataframes.
@@ -39,5 +41,14 @@ class Optimise:
                 df[col] = pd.to_datetime(df[col])
         return df
 
+
+def check_durations_sum_to_1(individuals, activities):
+    total_duration = [0.0] * len(individuals)  # Add up all the different activity durations
+    for activity in activities:
+        total_duration = total_duration + individuals.loc[:, f"{activity}{ColumnNames.ACTIVITY_DURATION}"]
+    if not (total_duration.apply(lambda x: round(x, 5)) == 1.0).all():
+        print("Some activity durations don't sum to 1", flush=True)
+        print(total_duration[total_duration != 1.0], flush=True)
+        raise Exception("Some activity durations don't sum to 1")
